pyproject.toml
[project]
name = "ai-agentic-development"
version = "0.1.0"
description = ""
authors = [
    {name = "iCode-Stack",email = "usman261295@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymongo (>=4.11.1,<5.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "uv (>=0.6.3,<0.7.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg-binary (>=3.2.5,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
